Split tests into unit vs integration tests
