+++ server/api/tests/integration/auth/test_login.py
@@ -1,7 +1,7 @@
 from django.urls import reverse
 from rest_framework.test import APITestCase
 from rest_framework import status
-from ...models import User
+from ....models import User
 
 
 class LoginTests(APITestCase):
