+++ server/api/tests/integration/auth/test_registration.py
@@ -1,7 +1,7 @@
 from django.urls import reverse
 from rest_framework.test import APITestCase
 from rest_framework import status
-from ...models import User
+from ....models import User
 
 
 class RegisterTests(APITestCase):
