+++ server/api/tests/integration/roles/test_role_crud.py
@@ -2,7 +2,7 @@ from django.urls import reverse
 from rest_framework.test import APITestCase
 from rest_framework import status
 from rest_framework_simplejwt.tokens import RefreshToken
-from ...models import (
+from ....models import (
     Workspace,
     WorkspaceMember,
     User,
