+++ server/api/tests/integration/shifts/test_shift_queries.py
@@ -3,7 +3,7 @@ from rest_framework.test import APITestCase
 from rest_framework import status
 from rest_framework_simplejwt.tokens import RefreshToken
 from datetime import datetime, timedelta, timezone
-from ...models import (
+from ....models import (
     Workspace,
     WorkspaceMember,
     User,
