+++ server/api/tests/integration/workspaces/test_workspace_members.py
@@ -2,7 +2,7 @@ from django.urls import reverse
 from rest_framework.test import APITestCase
 from rest_framework import status
 from rest_framework_simplejwt.tokens import RefreshToken
-from ...models import Workspace, WorkspaceMember, User, MemberPermissions
+from ....models import Workspace, WorkspaceMember, User, MemberPermissions
 
 
 class AddMemberTests(APITestCase):
